Добавил Database service с Postgres-базой и конвертами

+++ src/main/java/graf/server/Base/GameMechanics.java
@@ -3,7 +3,7 @@ package graf.server.Base;
 import java.util.Set;
 
 public interface GameMechanics extends Node, Runnable {
-    void registerGMSession(Set<Integer> userIDs);
+    void registerGMSession(Set<Long> userIDs);
 
-    boolean hasSession(Set<Integer> userIds);
+    boolean hasSession(Set<Long> userIds);
 }
